Moved blob processing to helper file
